Remove ordinal and add definable column types
In order to make the SA360 reports more usable and changeable
remove the 'ordinal' (which required you to re-count the columns
and their positions if anything changed in the report definition)
and instead use the actual column name in the parameters to decide
which to replace. The template processor checks the column value
for a match with the name, and replaces that with the parameter
value for the specific report. At the same time, it replaces the
column specifier withe desired SA360 type ("savedColumnName",
"columnName" etc) - the default being "savedColumnName".

Also a minor bugfix in install.sh to add the missing gmail api on
a fresh install.

Change-Id: I78946fdf1b255a11ebdd779e0409bd757a8ca8d8