Refactor and improve validator for SA360.
The SA360 class `sa360_v2.py` was getting too large and confusing. As a
consequence I've refactored it into two:
  sa360_dynamic.py for dynamic API reports
  sa360_web.py     for web downloadble queries
The knock on refactoring has been to remove the old file and all
references to it, making sure that the correct SA360 report types are
referred to where necessary.

Improving the SA360 validator was necessary to handle standard columns
(and, in the future, metrics and other column types). It has been
refactored to a common 'Validator' superclass, along with field list
holders for each of the report types.

Change-Id: I0bf8022952ff403e6b385f1434906e4210d6d197