GA360 implementation.
Add a report definition reader and parser for GA360 report request
objects. This will allow us to read and store the report objects
in Firestore, and replace parameters as appropriate.

This handles just the basics for "Metric Expression" type reports;
filters and segments in Metric Expression reports and other types
such as Histogram, Pivot and Cohort formats can be added later.

Change-Id: I78ff5e0478561fbf3d757ca649d981d07af4976c