Postprocessor bugfixes.
Removing the 'notify.topic' as no longer necessary stopped the
postprocessor from being invoked at all as this was the key used
to identify a postprocessor call. Replaced with the 'notify.message'
key instead.

Also the migration to Python 3.8 requires (on GCP) that the
environment be set in the cloud function. This affects the
job-monitor function only.

Change-Id: Ib1a105c3e264c986fb08e55789f5b928ea2c3f90