Bugfixes and improvements.
Add the 'manage_reports.py' class which can fetch and upload (create)
reports programmatically in DV360 and CM.

Update 'csv_helpers.py' to change the column names as we had an
edge-case where two columns were called
  'column name #' and
  'column name $',
which got sanitized to
  'column_name__' and
  'column_name__'
thus causing the report to fail. As a result, only [ ()] are
converted to '_', all other non-alphanumerics are converted to
the hex code of the character ie ' -> 0x27.

Update SA360 field validator to allow fetching of column names.

Update AbstractCredentials with better expiration/refresh file
handling.

Change-Id: Ia8fabb08516a7d6facfb89083a6287ab7d7f136c