Bugfixes for potential illegal column names.
In SA360, it is possible for the user to create a value beginning
with any character ('*' is legal, for example). When converted to a
column name by the 'csv_helpers.sanitize' function, this results in
a string beginning with '0x2a', which is an illegal column in
Big Query.

As a result, sanitize_string has been overloaded with two new
methods:
  'sanitize_title' which is used where the first character
  of the resultant string is immaterial, as in report2bq's table
  naming, it prepends the type and id of the report, so the table
  would be called 'sa360_hourly_depleted_0x2aSpecial' for example.
  'sanitize_column' where if the first character of the final
  string is in the set [0-9], an 'X' is prepended.

This has been necessary for backward compatibility as we need to
keep table names as is to avoid interrupting any current
downstream processes in existing instances of report2qb, but if
the column was illegal to begin with no table will have been
created yet so we can safely rename columns.

Change-Id: Icfa9ec20c8918af3c086bd4c7193e6d8cc47431e