Jinja2 error workaround.
Following the Jinja2 library upgrade to 3.1.0, any dependencies
just calling for 'jinja2' or 'jinja2>=x.x.x' will pull v3.1.x,
which is broken. The culprit here is Flask.

In order to get around this the workaround until Jinja2 is fixed
is to force 'jinja2==3.0.3', which is the last known 'good' version.

NOTE: I also tried 'jinja2!=3.1.0', knowing that v3.1.0 is broken
      and hoping that other 3.1.x versions after would have the
      issue sorted out, but 3.1.1 shows the same issue, hence the
      need to force jinja2 to a 3.0.x version.
Change-Id: I18212d000864ca37f2651630d7c58a67a7b5d18e